Register updates

See merge request fabrica-back-dev/gram-dahab!17

+++ app/Http/Controllers/Admin/ProductController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\ActiveProductRequest;
 use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
 use App\Http\Resources\ProductResource;
@@ -62,8 +63,10 @@ class ProductController extends Controller
     public function update(UpdateProductRequest $request , $id)
     {
         try{
-            $data = $request->validated();
             $model = $this->productRepository->find($id , []);
+            $this->authorize('update', $model);
+
+            $data = $request->validated();
             if($request->has('image')){
                 $fileName = $this->fileService->updateFile($data['image'] , $model);
                 $data['image'] = $fileName;
@@ -78,9 +81,26 @@ class ProductController extends Controller
     public function destroy($id)
     {
         try{
+            $model = $this->productRepository->find($id , []);
+            $this->authorize('delete', $model);
+
             $this->fileService->deleteFileFromUploads($id , $this->productRepository);
             $this->productRepository->delete($id);
             return response()->json(['message' => 'Product Deleted Successfully']);
+        }catch(\Exception $e){
+            return $e;
+        }
+    }
+
+    public function updateProduct(ActiveProductRequest $request , $id)
+    {
+        try{
+            $model = $this->productRepository->find($id , []);
+            $this->authorize('update', $model);
+
+            $data = $request->validated();
+            $this->productRepository->ActivateProductUpdate($model,$data);
+            return response()->json(['message' => 'Product Updated Successfully']);
         }catch(\Exception $e){
             return $e;
         }
